+++ admin/getting-started/src/input.css
@@ -1,0 +1,6 @@
+#wpcontent {
+    padding-left: 0;
+    @tailwind base;
+    @tailwind components;
+    @tailwind utilities;
+}
