+++ config/postcss.config.js
@@ -61,6 +61,7 @@ module.exports = {
 				},
 			},
 		} ),
+		require('tailwindcss')('./config/tailwind.config.js'),
 		require( 'autoprefixer' ),
 		require( 'postcss-color-function' ),
 	],
