+++ package.json
@@ -118,6 +118,7 @@
     "slick-carousel": "^1.8.1",
     "sshpk": "^1.16.1",
     "style-loader": "^0.19.1",
+    "tailwindcss": "^3.4.17",
     "webpack": "^4.44.1",
     "webpack-fix-style-only-entries": "^0.4.0"
   },
